requirements.txt
# Core libraries
ffmpeg-python==0.2.0

# Additional utilities
#numpy==1.21.0
#Pillow==8.3.1

# For progress bars (optional, but recommended for better user experience)
tqdm==4.62.0

# For handling audio metadata (optional, for advanced audio processing)
#mutagen==1.45.1

# For parallel processing (optional, for improved performance with large batches)
joblib==1.0.1

# For creating command-line interfaces (optional, but recommended for ease of use)
#click==8.0.1

# For handling environment variables (optional, but recommended for configuration management)
#python-dotenv==0.19.0

# For improved error handling and debugging (optional, but recommended)
#better-exceptions==0.3.3